+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/expense/ExpenseCancellationServiceImpl.java
@@ -29,6 +29,7 @@ import com.axelor.apps.hr.db.Expense;
 import com.axelor.apps.hr.db.HRConfig;
 import com.axelor.apps.hr.db.repo.ExpenseRepository;
 import com.axelor.apps.hr.exception.HumanResourceExceptionMessage;
+import com.axelor.apps.hr.service.KilometricExpenseService;
 import com.axelor.apps.hr.service.config.HRConfigService;
 import com.axelor.i18n.I18n;
 import com.axelor.message.db.Message;
@@ -46,6 +47,7 @@ public class ExpenseCancellationServiceImpl implements ExpenseCancellationServic
   protected ExpenseRepository expenseRepository;
   protected MoveRepository moveRepository;
   protected ExpenseFetchMoveService expenseFetchMoveService;
+  protected final KilometricExpenseService kilometricExpenseService;
 
   @Inject
   public ExpenseCancellationServiceImpl(
@@ -54,13 +56,15 @@ public class ExpenseCancellationServiceImpl implements ExpenseCancellationServic
       TemplateMessageService templateMessageService,
       ExpenseRepository expenseRepository,
       MoveRepository moveRepository,
-      ExpenseFetchMoveService expenseFetchMoveService) {
+      ExpenseFetchMoveService expenseFetchMoveService,
+      KilometricExpenseService kilometricExpenseService) {
     this.periodService = periodService;
     this.hrConfigService = hrConfigService;
     this.templateMessageService = templateMessageService;
     this.expenseRepository = expenseRepository;
     this.moveRepository = moveRepository;
     this.expenseFetchMoveService = expenseFetchMoveService;
+    this.kilometricExpenseService = kilometricExpenseService;
   }
 
   @Override
@@ -70,6 +74,7 @@ public class ExpenseCancellationServiceImpl implements ExpenseCancellationServic
     Move move = expenseFetchMoveService.getExpenseMove(expense);
     if (move == null) {
       expense.setStatusSelect(ExpenseRepository.STATUS_CANCELED);
+      kilometricExpenseService.updateExpenseLineKilometricLog(expense);
       expenseRepository.save(expense);
       return;
     }
@@ -78,6 +83,7 @@ public class ExpenseCancellationServiceImpl implements ExpenseCancellationServic
       moveRepository.remove(move);
       expense.setVentilated(false);
       expense.setStatusSelect(ExpenseRepository.STATUS_CANCELED);
+      kilometricExpenseService.updateExpenseLineKilometricLog(expense);
     } catch (Exception e) {
       throw new AxelorException(
           e,
