+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/web/expense/ExpenseController.java
@@ -55,6 +55,7 @@ import com.axelor.apps.hr.db.repo.ExpenseRepository;
 import com.axelor.apps.hr.exception.HumanResourceExceptionMessage;
 import com.axelor.apps.hr.service.HRMenuTagService;
 import com.axelor.apps.hr.service.HRMenuValidateService;
+import com.axelor.apps.hr.service.KilometricExpenseService;
 import com.axelor.apps.hr.service.KilometricService;
 import com.axelor.apps.hr.service.app.AppHumanResourceService;
 import com.axelor.apps.hr.service.expense.ExpenseAnalyticService;
@@ -596,7 +597,8 @@ public class ExpenseController {
 
     BigDecimal amount = BigDecimal.ZERO;
     try {
-      amount = Beans.get(KilometricService.class).computeKilometricExpense(expenseLine, employee);
+      amount =
+          Beans.get(KilometricExpenseService.class).computeKilometricExpense(expenseLine, employee);
     } catch (Exception e) {
       TraceBackService.trace(response, e);
     }
@@ -730,7 +732,8 @@ public class ExpenseController {
             AuthUtils.getUser().getName());
       }
 
-      amount = kilometricService.computeKilometricExpense(expenseLine, employee);
+      amount =
+          Beans.get(KilometricExpenseService.class).computeKilometricExpense(expenseLine, employee);
       response.setValue("totalAmount", amount);
       response.setValue("untaxedAmount", amount);
